+++ src/cuda/ddm_cuda.cc
@@ -31,7 +31,7 @@ py::array_t<double> ddm_diff_cuda(py::array_t<T, py::array::c_style> img_seq,
     auto p_img_seq = img_seq.data();                // get input data
 
     // Check host memory
-    chk_host_mem_diff(nx, ny, length, lags);
+    chk_host_mem_diff(nx, ny, length, lags.size());
 
     // Check device memory and optimize
     unsigned long long num_fft2, num_chunks, num_fullshift;
@@ -59,7 +59,8 @@ py::array_t<double> ddm_diff_cuda(py::array_t<T, py::array::c_style> img_seq,
       double [the input needs to be twice as large]
      */
     unsigned long long _nx = nx / 2 + 1;
-    py::array_t<double> out = py::array_t<double>(2 * _nx * ny * length);
+    unsigned long long dim_t = max(length, (unsigned long long)(lags.size() + 2));
+    py::array_t<double> out = py::array_t<double>(2 * _nx * ny * dim_t);
     auto p_out = out.mutable_data();
 
     // ***Transfer data to GPU and compute fft2
@@ -86,7 +87,7 @@ py::array_t<double> ddm_diff_cuda(py::array_t<T, py::array::c_style> img_seq,
     // ***Convert raw output to full and shifted image structure function
     // Convert raw output to full and shifted image structure function
     make_full_shift(p_out,
-                    lags,
+                    lags.size() + 2,
                     nx,
                     ny,
                     num_fullshift,
@@ -95,7 +96,10 @@ py::array_t<double> ddm_diff_cuda(py::array_t<T, py::array::c_style> img_seq,
     // ***Resize output
     // the full size of the image structure function is
     // nx * ny * #(lags)
-    out.resize({(unsigned long long)(lags.size()), ny, nx});
+    out.resize({(unsigned long long)(lags.size() + 2), ny, nx});
+
+    // release pointer to output array
+    p_out = NULL;
 
     // ***Return result to python
     return out;
@@ -122,7 +126,7 @@ py::array_t<double> ddm_fft_cuda(py::array_t<T, py::array::c_style> img_seq,
     auto p_img_seq = img_seq.data();                // get input data
 
     // Check host memory
-    chk_host_mem_fft(nx, ny, length);
+    chk_host_mem_fft(nx, ny, length, lags.size());
 
     // Check device memory and optimize
     unsigned long long num_fft2, num_chunks, num_fullshift;
@@ -152,7 +156,8 @@ py::array_t<double> ddm_fft_cuda(py::array_t<T, py::array::c_style> img_seq,
       doubles [the input needs to be twice as large]
      */
     unsigned long long _nx = nx / 2 + 1;
-    py::array_t<double> out = py::array_t<double>(2 * _nx * ny * length);
+    unsigned long long dim_t = max(length, (unsigned long long)(lags.size() + 2));
+    py::array_t<double> out = py::array_t<double>(2 * _nx * ny * dim_t);
     auto p_out = out.mutable_data();
 
     // ***Transfer data to GPU and compute fft2
@@ -180,7 +185,7 @@ py::array_t<double> ddm_fft_cuda(py::array_t<T, py::array::c_style> img_seq,
 
     // ***Convert raw output to full and shifted image structure function
     make_full_shift(p_out,
-                    lags,
+                    lags.size() + 2,
                     nx,
                     ny,
                     num_fullshift,
@@ -189,7 +194,10 @@ py::array_t<double> ddm_fft_cuda(py::array_t<T, py::array::c_style> img_seq,
     // ***Resize output
     // the full size of the image structure function is
     // nx * ny * #(lags)
-    out.resize({(unsigned long long)(lags.size()), ny, nx});
+    out.resize({(unsigned long long)(lags.size() + 2), ny, nx});
+
+    // release pointer to output array
+    p_out = NULL;
 
     // ***Return result to python
     return out;
@@ -224,22 +232,22 @@ void export_ddm_cuda(py::module &m)
     // m.def("get_device_fft2_mem", &get_device_fft2_mem);
     // m.def("get_device_fft_mem", &get_device_fft_mem);
     // Leave function export in this order!
-    m.def("ddm_diff_cuda", &ddm_diff_cuda<uint8_t>);
-    m.def("ddm_diff_cuda", &ddm_diff_cuda<int16_t>);
-    m.def("ddm_diff_cuda", &ddm_diff_cuda<uint16_t>);
-    m.def("ddm_diff_cuda", &ddm_diff_cuda<int32_t>);
-    m.def("ddm_diff_cuda", &ddm_diff_cuda<uint32_t>);
-    m.def("ddm_diff_cuda", &ddm_diff_cuda<int64_t>);
-    m.def("ddm_diff_cuda", &ddm_diff_cuda<uint64_t>);
-    m.def("ddm_diff_cuda", &ddm_diff_cuda<float>);
-    m.def("ddm_diff_cuda", &ddm_diff_cuda<double>);
-    m.def("ddm_fft_cuda", &ddm_fft_cuda<uint8_t>);
-    m.def("ddm_fft_cuda", &ddm_fft_cuda<int16_t>);
-    m.def("ddm_fft_cuda", &ddm_fft_cuda<uint16_t>);
-    m.def("ddm_fft_cuda", &ddm_fft_cuda<int32_t>);
-    m.def("ddm_fft_cuda", &ddm_fft_cuda<uint32_t>);
-    m.def("ddm_fft_cuda", &ddm_fft_cuda<int64_t>);
-    m.def("ddm_fft_cuda", &ddm_fft_cuda<uint64_t>);
-    m.def("ddm_fft_cuda", &ddm_fft_cuda<float>);
-    m.def("ddm_fft_cuda", &ddm_fft_cuda<double>);
+    m.def("ddm_diff_cuda", &ddm_diff_cuda<uint8_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_diff_cuda", &ddm_diff_cuda<int16_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_diff_cuda", &ddm_diff_cuda<uint16_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_diff_cuda", &ddm_diff_cuda<int32_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_diff_cuda", &ddm_diff_cuda<uint32_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_diff_cuda", &ddm_diff_cuda<int64_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_diff_cuda", &ddm_diff_cuda<uint64_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_diff_cuda", &ddm_diff_cuda<float>, py::return_value_policy::take_ownership);
+    m.def("ddm_diff_cuda", &ddm_diff_cuda<double>, py::return_value_policy::take_ownership);
+    m.def("ddm_fft_cuda", &ddm_fft_cuda<uint8_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_fft_cuda", &ddm_fft_cuda<int16_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_fft_cuda", &ddm_fft_cuda<uint16_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_fft_cuda", &ddm_fft_cuda<int32_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_fft_cuda", &ddm_fft_cuda<uint32_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_fft_cuda", &ddm_fft_cuda<int64_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_fft_cuda", &ddm_fft_cuda<uint64_t>, py::return_value_policy::take_ownership);
+    m.def("ddm_fft_cuda", &ddm_fft_cuda<float>, py::return_value_policy::take_ownership);
+    m.def("ddm_fft_cuda", &ddm_fft_cuda<double>, py::return_value_policy::take_ownership);
 }