+++ src/python/make_install_setup.py
@@ -8,7 +8,7 @@ setup(
     name = "fastddm",
     use_scm_version = {
         "root": "${CMAKE_SOURCE_DIR}",
-        "fallback_version": "0.1.1",
+        "fallback_version": "0.1.2",
     },
     setup_requires = ["setuptools_scm"],
     packages = ["fastddm"],